Extract reuse methods
