+++ app/workers/shared_code/helpers.py
@@ -1,6 +1,6 @@
 import json
 import os
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Literal, Optional, Tuple
 
 import azure.functions as func
 
@@ -257,3 +257,32 @@ def add_vocabulary_id_to_entries(
                 field_name = field["name"]
                 vocab_id = vocab[table_name].get(field_name)
         value["vocabulary_id"] = vocab_id
+
+
+def create_concept(
+    concept_id: str,
+    object_id: str,
+    content_type: Literal[15, 17] = 17,
+    creation_type: Literal["V", "R"] = "V",
+) -> Dict[str, Any]:
+    """
+    Creates a new Concept dict.
+
+    Args:
+        concept_id (str): The Id of the Concept to create.
+        object_id (str): The Object Id of the Concept to create.
+        content_type (Literal[15, 17], optional): The Content Type Id of the Concept.
+        creation_type (Literal["R", "V"], optional): The Creation Type value of the Concept.
+
+    Returns:
+        Dict[str, Any]: A Concept as a dictionary.
+
+    TODO: we should query `content_type` from the API, when this is fixed:
+    https://github.com/Health-Informatics-UoN/CaRROT-Mapper/issues/637
+    """
+    return {
+        "concept": concept_id,
+        "object_id": object_id,
+        "content_type": content_type,
+        "creation_type": creation_type,
+    }
